Update README.md
Using wget downloads a file called xt-code with no extension so the following command to chmod and run don't work since it a a different file name, I just changed the second command to be xt-code but you could also add a new command to rename the file perhaps? You can do a different fix if you don't want to merge this.